Moved successfully into controller - reduced route
