+++ app/routes/application.js
@@ -1,5 +1,4 @@
 import Route from '@ember/routing/route';
-import { action } from '@ember-decorators/object';
 
 export default class ApplicationRoute extends Route {
   beforeModel(transition) {
@@ -7,10 +6,5 @@ export default class ApplicationRoute extends Route {
     if (transition.intent.url === '/') {
       this.transitionTo('show-geography');
     }
-  }
-
-  @action
-  refreshModel() {
-    this.refresh();
   }
 }
